# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Epistemic Transparency \u0026 Event Lineage System API\nAn append-only event store with hash-chaining for AI decision tracking",
        "title": "Lineage API",
        "contact": {
            "name": "Lineage Support",
            "url": "https://github.com/vkrishna03/Lineage"
        },
        "license": {
            "name": "Elastic License 2.0",
            "url": "https://www.elastic.co/licensing/elastic-license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/event-types": {
            "get": {
                "description": "List all active event types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event-types"
                ],
                "summary": "List event types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/eventtype.ListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new event type with optional JSON schema for payload validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event-types"
                ],
                "summary": "Create an event type",
                "parameters": [
                    {
                        "description": "Event type creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/eventtype.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/eventtype.EventTypeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/event-types/{id}": {
            "get": {
                "description": "Get an event type by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event-types"
                ],
                "summary": "Get an event type",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Event Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/eventtype.EventTypeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/scopes": {
            "get": {
                "description": "List all scopes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scopes"
                ],
                "summary": "List scopes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scope.ListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new scope for grouping events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scopes"
                ],
                "summary": "Create a scope",
                "parameters": [
                    {
                        "description": "Scope creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scope.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/scope.ScopeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/scopes/{id}": {
            "get": {
                "description": "Get a scope by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scopes"
                ],
                "summary": "Get a scope",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Scope ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scope.ScopeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "eventtype.CreateRequest": {
            "type": "object",
            "required": [
                "name",
                "version"
            ],
            "properties": {
                "allowed_intents": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "exploration",
                        "suggestion",
                        "assertion",
                        "decision",
                        "execution"
                    ]
                },
                "description": {
                    "type": "string",
                    "example": "A decision event type"
                },
                "name": {
                    "type": "string",
                    "example": "decision"
                },
                "payload_schema": {
                    "type": "object"
                },
                "version": {
                    "type": "string",
                    "example": "1.0"
                }
            }
        },
        "eventtype.EventTypeResponse": {
            "description": "Event type data returned by the API",
            "type": "object",
            "properties": {
                "allowed_intents": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "exploration",
                        "suggestion",
                        "assertion",
                        "decision",
                        "execution"
                    ]
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "A decision event type"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "decision"
                },
                "payload_schema": {
                    "type": "object"
                },
                "version": {
                    "type": "string",
                    "example": "1.0"
                }
            }
        },
        "eventtype.ListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "event_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/eventtype.EventTypeResponse"
                    }
                }
            }
        },
...
        "scope.CreateRequest": {
            "type": "object",
            "required": [
                "project"
            ],
            "properties": {
                "domain": {
                    "type": "string",
                    "example": "payments"
                },
                "environment": {
                    "type": "string",
                    "example": "production"
                },
                "project": {
                    "type": "string",
                    "example": "my-project"
                }
            }
        },
        "scope.ListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/scope.ScopeResponse"
                    }
                }
            }
        },
        "scope.ScopeResponse": {
            "description": "Scope data returned by the API",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "domain": {
                    "type": "string",
                    "example": "payments"
                },
                "environment": {
                    "type": "string",
                    "example": "production"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "project": {
                    "type": "string",
                    "example": "my-project"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}